+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
    + Aggregates are not supported
  - #2967, Add `Proxy-Status` header for better error response - @taimoorzaeem
  - #4016, Add `Content-Length` response header - @laurenceisla
+ - #3410, Add `metrics=timings` preference to Prefer header - @taimoorzaeem
 
 ### Fixed